+++ odoo/local-src/specific_stock/tests/__init__.py
@@ -2,3 +2,4 @@
 
 from . import test_picking_transfer
 from . import test_product_category_expired_times
+from . import test_lot_checksum