+++ odoo/local-src/stock_expired/models/stock_picking.py
@@ -8,4 +8,5 @@ from odoo import models, fields
 class StockPicking(models.Model):
     _inherit = 'stock.picking'
 
-    to_process_quant_expired = fields.Boolean()
+    to_process_quant_expired = fields.Boolean(
+        'Bypass restriction on expired quants')